requirements.txt
Django==2.2.6
django-js-asset==2.0.0
django-mptt==0.14.0
Pillow==9.2.0
pytz==2022.5
sqlparse==0.4.3
